fixed delete ClientMessage

+++ src/xlib_window_system.rs
@@ -150,7 +150,7 @@ impl XlibWindowSystem {
         msg.display = self.display;
         msg.window = window;
         msg.message_type = XInternAtom(self.display, "WM_PROTOCOLS".to_c_str().as_mut_ptr(), 1);
-        msg.data = [XInternAtom(self.display, "WM_DELETE_WINDOW".to_c_str().as_mut_ptr(), 1) as u32, 0, 0, 0, 0];
+        msg.data = [XInternAtom(self.display, "WM_DELETE_WINDOW".to_c_str().as_mut_ptr(), 1), 0, 0, 0, 0];
 
         XSendEvent(self.display, window, 0, 0, transmute(&msg));
       } else {
